requirements.txt
pandas==2.1.4
apache-airflow-providers-google==15.1.0
dbt-bigquery==1.9.0

